data_stopwords_ancient {stopwords}	R Documentation

stopword lists for ancient languages

Description

Stopword lists for ancient Greek and Latin. These lists are far more extensive than the Perseus lists for ancient Greek and Latin from the Perseus Digital Library.

Format

An object of class list of length 2.

Details

As there is no 2-letter code for ancient Greek in ISO-639-1, we use "grc" to denote Greek (as per ISO-639-3).

Usage

stopwords(language = "grc", source = "ancient")

stopwords(language = "la", source = "ancient")

Source

Aurélien Berra, Ancient Greek and Latin stopwords, doi: 10.5281/zenodo.1165205. See https://github.com/aurelberra/stopwords/blob/master/rationale.md.

See Also

data_stopwords_perseus

[Package stopwords version 2.2 Index]