stopword lists including parts-of-speech

Stopword lists that include specific parts of speech, maintained by Kohei Watanabe.

Format

An object of class list of length 6.

Source

The English version was adopted from the Snowball collection, and then extended and translated into other languages by contributors. Names of contributors are in the header of the original YAML files.

Details

These are multi-level lists, in the original data. If you wish to use them as lists, please access the data object directly.

Usage

stopwords(language = "en", source = "marimo")

Examples

# access English pronouns directly
stopwords::data_stopwords_marimo$en$pronoun

#> $basic
#>  [1] "i"          "me"         "myself"     "we"         "ours"      
#>  [6] "ourselves"  "you"        "yours"      "yourself"   "yourselves"
#> [11] "he"         "him"        "himself"    "she"        "hers"      
#> [16] "herself"    "it"         "itself"     "they"       "them"      
#> [21] "theirs"     "themselves" "this"       "that"       "these"     
#> [26] "those"     
#> 
#> $possessive
#> [1] "my"    "our"   "your"  "his"   "her"   "its"   "their"
#> 
#> $interrogative
#> [1] "what"  "which" "who"   "whom"  "whose" "when"  "where" "why"   "how"  
#> 
#> $contraction
#>  [1] "i'm"     "you're"  "he's"    "she's"   "it's"    "we're"   "they're"
#>  [8] "i've"    "you've"  "we've"   "they've" "i'd"     "you'd"   "he'd"   
#> [15] "she'd"   "we'd"    "they'd"  "i'll"    "you'll"  "he'll"   "she'll" 
#> [22] "we'll"   "they'll"
#>