data_stopwords_misc {stopwords}	R Documentation

miscellaneous stopword lists

Description

Other, miscellaneous stopword lists.

Format

An object of class list of length 5.

Usage

stopwords(language, source = "misc")

Source

The Arabic stopwords come from https://sites.google.com/site/kevinbouge/stopwords-lists.

The Catalan stopwords come from http://latel.upf.edu/morgana/altres/pub/ca_stop.htm.

The Greek stopwords were supplied by Carsten Schwemmer (see https://github.com/quanteda/quanteda/issues/282).

The Gujarati stopwords are taken from https://github.com/gujarati-ir/Gujarati-Stop-Words and modified by Chandrakant Bhogayata.

The Chinese stopwords are taken from the Baidu stopword list (see http://www.baiduguide.com/baidu-stopwords/).

[Package stopwords version 2.2 Index]