stopword lists from the SMART system

The stopword lists based on the SMART (System for the Mechanical Analysis and Retrieval of Text) Information Retrieval System, an information retrieval system developed at Cornell University in the 1960s.

Format

An object of class list of length 1.

Source

The English stopword list is taken from the online appendix 11 of Lewis et. al. (2004).

Usage

stopwords(language = "en", source = "smart")

References

Lewis, David D., et al. (2004) "Rcv1: A new benchmark collection for text categorization research." Journal of machine learning research 5: 361-397.