snowball stopword list

snowball stopword list

Format

An object of class list of length 15.

Source

The main stopword lists are taken from the Snowball stemmer project in different languages (see http://snowballstem.org/projects.html).

The stopword lists can be found in http://snowball.tartarus.org/dist/snowball_all.tgz.

Details

Provides stopword lists in multiple languages, based on the Snowball stemmer's word lists.

Usage

stopwords(language, source = "snowball")

See also