Use stopwords in your package

Sets up your package to import and re-export the stopwords() function in your package. This function requires the use of the roxygen2 package.

Adds stopwords`` to "Imports" in DESCRIPTION`.

Imports stopwords(), which is necessary for internal use.

Exports stopwords(), if export = TRUE, which is necessary to make the function available to the users of your package.

use_stopwords(export = TRUE)

Arguments

export
If TRUE, the file R/stopwords.R is added, which provides the roxygen template to import and re-export stopwords(). If FALSE, the necessary roxygen directive is added, if possible, or otherwise instructions are given.

Examples

if (FALSE) {
use_stopwords()
}