data_stopwords_marimo {stopwords}	R Documentation

stopword lists including parts-of-speech

Description

Stopword lists that include specific parts of speech, maintained by Kohei Watanabe.

Format

An object of class list of length 8.

Details

These are multi-level lists, in the original data. If you wish to use them as lists, please access the data object directly.

Usage

stopwords(language = "en", source = "marimo")

Source

The English version was adopted from the Snowball collection, and then extended and translated into other languages by contributors. Names of contributors are in the header of the original YAML files.

Examples

# access English pronouns directly
stopwords::data_stopwords_marimo$en$pronoun

[Package stopwords version 2.2 Index]