data_stopwords_snowball {stopwords}	R Documentation

snowball stopword list

Description

snowball stopword list

Format

An object of class list of length 15.

Details

Provides stopword lists in multiple languages, based on the Snowball stemmer's word lists.

Usage

stopwords(language, source = "snowball")

Source

The main stopword lists are taken from the Snowball stemmer project in different languages (see https://snowballstem.org/projects.html).

The stopword lists can be found in http://snowball.tartarus.org/dist/snowball_all.tgz.

See Also

stopwords()

[Package stopwords version 2.2 Index]