Collection of stopwords in multiple languages

This function returns character vectors of stopwords for different languages, using the ISO-639-1 language codes, and allows for different sources of stopwords to be defined.

The default source is the Snowball() stopwords collection but other() sources are also available.

stopwords(language = "en", source = "snowball", simplify = TRUE)

Arguments

language
specify language of stopwords by ISO 639-1 code

source
specify a stopwords source. To list the currently available options, use stopwords_getsources().

simplify
logical; if TRUE return a simple vector, if FALSE return a list if the original word list was nested

Value

a character vector containing the stopwords, or a list of characters simplify = FALSE

Details

The language codes for each stopword list use the two-letter ISO code from https://en.wikipedia.org/wiki/List_of_ISO_639-1_codes. For backwards compatibility, the full English names of the stopwords from the quanteda package may also be used, although these are deprecated.

Examples

stopwords("en")

#>   [1] "i"          "me"         "my"         "myself"     "we"        
#>   [6] "our"        "ours"       "ourselves"  "you"        "your"      
#>  [11] "yours"      "yourself"   "yourselves" "he"         "him"       
#>  [16] "his"        "himself"    "she"        "her"        "hers"      
#>  [21] "herself"    "it"         "its"        "itself"     "they"      
#>  [26] "them"       "their"      "theirs"     "themselves" "what"      
#>  [31] "which"      "who"        "whom"       "this"       "that"      
#>  [36] "these"      "those"      "am"         "is"         "are"       
#>  [41] "was"        "were"       "be"         "been"       "being"     
#>  [46] "have"       "has"        "had"        "having"     "do"        
#>  [51] "does"       "did"        "doing"      "would"      "should"    
#>  [56] "could"      "ought"      "i'm"        "you're"     "he's"      
#>  [61] "she's"      "it's"       "we're"      "they're"    "i've"      
#>  [66] "you've"     "we've"      "they've"    "i'd"        "you'd"     
#>  [71] "he'd"       "she'd"      "we'd"       "they'd"     "i'll"      
#>  [76] "you'll"     "he'll"      "she'll"     "we'll"      "they'll"   
#>  [81] "isn't"      "aren't"     "wasn't"     "weren't"    "hasn't"    
#>  [86] "haven't"    "hadn't"     "doesn't"    "don't"      "didn't"    
#>  [91] "won't"      "wouldn't"   "shan't"     "shouldn't"  "can't"     
#>  [96] "cannot"     "couldn't"   "mustn't"    "let's"      "that's"    
#> [101] "who's"      "what's"     "here's"     "there's"    "when's"    
#> [106] "where's"    "why's"      "how's"      "a"          "an"        
#> [111] "the"        "and"        "but"        "if"         "or"        
#> [116] "because"    "as"         "until"      "while"      "of"        
#> [121] "at"         "by"         "for"        "with"       "about"     
#> [126] "against"    "between"    "into"       "through"    "during"    
#> [131] "before"     "after"      "above"      "below"      "to"        
#> [136] "from"       "up"         "down"       "in"         "out"       
#> [141] "on"         "off"        "over"       "under"      "again"     
#> [146] "further"    "then"       "once"       "here"       "there"     
#> [151] "when"       "where"      "why"        "how"        "all"       
#> [156] "any"        "both"       "each"       "few"        "more"      
#> [161] "most"       "other"      "some"       "such"       "no"        
#> [166] "nor"        "not"        "only"       "own"        "same"      
#> [171] "so"         "than"       "too"        "very"       "will"      

stopwords("de")

#>   [1] "aber"      "alle"      "allem"     "allen"     "aller"     "alles"    
#>   [7] "als"       "also"      "am"        "an"        "ander"     "andere"   
#>  [13] "anderem"   "anderen"   "anderer"   "anderes"   "anderm"    "andern"   
#>  [19] "anderr"    "anders"    "auch"      "auf"       "aus"       "bei"      
#>  [25] "bin"       "bis"       "bist"      "da"        "damit"     "dann"     
#>  [31] "der"       "den"       "des"       "dem"       "die"       "das"      
#>  [37] "daß"       "derselbe"  "derselben" "denselben" "desselben" "demselben"
#>  [43] "dieselbe"  "dieselben" "dasselbe"  "dazu"      "dein"      "deine"    
#>  [49] "deinem"    "deinen"    "deiner"    "deines"    "denn"      "derer"    
#>  [55] "dessen"    "dich"      "dir"       "du"        "dies"      "diese"    
#>  [61] "diesem"    "diesen"    "dieser"    "dieses"    "doch"      "dort"     
#>  [67] "durch"     "ein"       "eine"      "einem"     "einen"     "einer"    
#>  [73] "eines"     "einig"     "einige"    "einigem"   "einigen"   "einiger"  
#>  [79] "einiges"   "einmal"    "er"        "ihn"       "ihm"       "es"       
#>  [85] "etwas"     "euer"      "eure"      "eurem"     "euren"     "eurer"    
#>  [91] "eures"     "für"       "gegen"     "gewesen"   "hab"       "habe"     
#>  [97] "haben"     "hat"       "hatte"     "hatten"    "hier"      "hin"      
#> [103] "hinter"    "ich"       "mich"      "mir"       "ihr"       "ihre"     
#> [109] "ihrem"     "ihren"     "ihrer"     "ihres"     "euch"      "im"       
#> [115] "in"        "indem"     "ins"       "ist"       "jede"      "jedem"    
#> [121] "jeden"     "jeder"     "jedes"     "jene"      "jenem"     "jenen"    
#> [127] "jener"     "jenes"     "jetzt"     "kann"      "kein"      "keine"    
#> [133] "keinem"    "keinen"    "keiner"    "keines"    "können"    "könnte"   
#> [139] "machen"    "man"       "manche"    "manchem"   "manchen"   "mancher"  
#> [145] "manches"   "mein"      "meine"     "meinem"    "meinen"    "meiner"   
#> [151] "meines"    "mit"       "muss"      "musste"    "nach"      "nicht"    
#> [157] "nichts"    "noch"      "nun"       "nur"       "ob"        "oder"     
#> [163] "ohne"      "sehr"      "sein"      "seine"     "seinem"    "seinen"   
#> [169] "seiner"    "seines"    "selbst"    "sich"      "sie"       "ihnen"    
#> [175] "sind"      "so"        "solche"    "solchem"   "solchen"   "solcher"  
#> [181] "solches"   "soll"      "sollte"    "sondern"   "sonst"     "über"     
#> [187] "um"        "und"       "uns"       "unse"      "unsem"     "unsen"    
#> [193] "unser"     "unses"     "unter"     "viel"      "vom"       "von"      
#> [199] "vor"       "während"   "war"       "waren"     "warst"     "was"      
#> [205] "weg"       "weil"      "weiter"    "welche"    "welchem"   "welchen"  
#> [211] "welcher"   "welches"   "wenn"      "werde"     "werden"    "wie"      
#> [217] "wieder"    "will"      "wir"       "wird"      "wirst"     "wo"       
#> [223] "wollen"    "wollte"    "würde"     "würden"    "zu"        "zum"      
#> [229] "zur"       "zwar"      "zwischen"