stopwords: one-stop shopping for stopwords in R

Provides a stopwords() function to return character vectors of stopwords for different languages, using the ISO-639-1 language codes, and allows for different sources of stopwords to be defined.

Currently available sources

snowball

The Snowball stopword lists sources for multiple languages. Most of these have been ported from the quanteda stopword lists (in versions <1.0 of that package).

stopwords-iso

The collection taken from https://github.com/stopwords-iso/stopwords-iso/.

smart

The English-language stopword list from the SMART information retrieval system.

misc

A few additional stopword lists, including the non-Snowball word lists from quanteda versions < 1.0.

marimo

Stopword lists compiled by Kohei Watanabe.

Author

Kenneth Benoit, David Muhr, and Kohei Watanabe